fix: 미들웨어 주체변동 => 클라이언트 사이드에서 컨트롤
